Extract project index concurrency helpers
