+++ src/plugin/src/options/identifier-case.js
@@ -5,10 +5,8 @@
 // single source of truth.
 
 import { capitalize } from "../../../shared/string-utils.js";
-import {
-    DEFAULT_MAX_PROJECT_INDEX_CACHE_SIZE,
-    getDefaultProjectIndexGmlConcurrency
-} from "../project-index/index.js";
+import { DEFAULT_MAX_PROJECT_INDEX_CACHE_SIZE } from "../project-index/cache.js";
+import { getDefaultProjectIndexGmlConcurrency } from "../project-index/concurrency.js";
 import { normalizeStringList } from "./option-utils.js";
 
 export const DEFAULT_IDENTIFIER_CASE_OPTION_STORE_MAX_ENTRIES = 128;
