+++ src/plugin/src/project-index/index.js
@@ -13,6 +13,7 @@ import {
     buildFileLocationKey
 } from "../../../shared/location-keys.js";
 import { getDefaultProjectIndexParser } from "./gml-parser-facade.js";
+import { clampConcurrency } from "./concurrency.js";
 import {
     PROJECT_MANIFEST_EXTENSION,
     isProjectManifestPath
@@ -31,61 +32,6 @@ import {
 import { throwIfAborted } from "../../../shared/abort-utils.js";
 
 const defaultProjectIndexParser = getDefaultProjectIndexParser();
-
-const PROJECT_INDEX_GML_CONCURRENCY_ENV_VAR = "GML_PROJECT_INDEX_CONCURRENCY";
-const PROJECT_INDEX_GML_CONCURRENCY_BASELINE = 4;
-
-let configuredDefaultProjectIndexGmlConcurrency =
-    PROJECT_INDEX_GML_CONCURRENCY_BASELINE;
-
-function getDefaultProjectIndexGmlConcurrency() {
-    return configuredDefaultProjectIndexGmlConcurrency;
-}
-
-function normalizeDefaultProjectIndexConcurrencyInput(value) {
-    if (value === undefined || value === null) {
-        return null;
-    }
-
-    const trimmed =
-        typeof value === "string" ? value.trim() : value;
-    if (trimmed === "") {
-        return null;
-    }
-
-    const numeric = Number(trimmed);
-    return Number.isFinite(numeric) ? numeric : null;
-}
-
-function setDefaultProjectIndexGmlConcurrency(concurrency) {
-    const normalized = normalizeDefaultProjectIndexConcurrencyInput(concurrency);
-
-    if (normalized === null) {
-        configuredDefaultProjectIndexGmlConcurrency =
-            PROJECT_INDEX_GML_CONCURRENCY_BASELINE;
-        return configuredDefaultProjectIndexGmlConcurrency;
-    }
-
-    configuredDefaultProjectIndexGmlConcurrency = clampConcurrency(normalized, {
-        fallback: PROJECT_INDEX_GML_CONCURRENCY_BASELINE
-    });
-
-    return configuredDefaultProjectIndexGmlConcurrency;
-}
-
-function applyProjectIndexConcurrencyEnvOverride(env = process?.env) {
-    const rawValue = env?.[PROJECT_INDEX_GML_CONCURRENCY_ENV_VAR];
-    if (rawValue === undefined) {
-        return;
-    }
-
-    setDefaultProjectIndexGmlConcurrency(rawValue);
-}
-
-applyProjectIndexConcurrencyEnvOverride();
-
-const DEFAULT_PROJECT_INDEX_GML_CONCURRENCY =
-    getDefaultProjectIndexGmlConcurrency();
 
 const PARSER_FACADE_OPTION_KEYS = [
     "identifierCaseProjectIndexParserFacade",
@@ -296,7 +242,7 @@ export {
     setDefaultProjectIndexGmlConcurrency,
     PROJECT_INDEX_GML_CONCURRENCY_ENV_VAR,
     PROJECT_INDEX_GML_CONCURRENCY_BASELINE
-};
+} from "./concurrency.js";
 
 const GML_IDENTIFIER_FILE_PATH = fileURLToPath(
     new URL("../../../../resources/gml-identifiers.json", import.meta.url)
@@ -2155,21 +2101,6 @@ function cloneAssetReference(reference) {
         targetName: reference.targetName ?? null,
         targetResourceType: reference.targetResourceType ?? null
     };
-}
-
-function clampConcurrency(
-    value,
-    { min = 1, max = 16, fallback = getDefaultProjectIndexGmlConcurrency() } = {}
-) {
-    const candidate = value ?? fallback;
-    const numeric = Number(candidate);
-    if (!Number.isFinite(numeric) || numeric < min) {
-        return min;
-    }
-    if (numeric > max) {
-        return max;
-    }
-    return numeric;
 }
 
 async function processWithConcurrency(items, limit, worker, options = {}) {
